+++ src/snowflake/snowpark/modin/plugin/docstrings/series.py
@@ -2316,6 +2316,32 @@ class Series(BasePandasDataset):
         Make plot of Series.
         """
 
+    def pop():
+        """
+        Return item and drops from series. Raise KeyError if not found.
+
+        Parameters
+        ----------
+        item : label
+            Index of the element that needs to be removed.
+
+        Returns
+        -------
+            Value that is popped from series.
+
+        Examples
+        --------
+        >>> ser = pd.Series([1, 2, 3])
+
+        >>> ser.pop(0)
+        1
+
+        >>> ser
+        1    2
+        2    3
+        dtype: int64
+        """
+
     @_create_operator_docstring(pandas.core.series.Series.pow, overwrite_existing=True)
     def pow():
         pass