+++ CHANGELOG.md
@@ -81,6 +81,7 @@
 - Added support for `Series.between`.
 - Added support for `include_groups=False` in `DataFrameGroupBy.apply`.
 - Added support for `expand=True` in `Series.str.split`.
+- Added support for `DataFrame.pop` and `Series.pop`.
 
 #### Bug Fixes